requirements.txt
absl-py==0.7.1
astor==0.8.0
cycler==0.10.0
gast==0.2.2
grpcio==1.16.0
h5py==1.10.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
numpy==1.16.5
opencv-python==4.1.1.26
protobuf==3.8.0
pyparsing==2.4.2
python-dateutil==2.8.0
scipy==1.3.1
Shapely==1.6.4
six==1.12.0
scikit-image==0.16.2
scikit-learn==0.21.3
tensorboard==1.12.2
tensorflow-gpu==1.12.0
termcolor==1.1.0
tqdm==4.42.1
Werkzeug==0.15.5
